refactor: (BREAKING) Replace Event.relationshipType and Event.relationshipEventId with Event.inReplyToEventId() for replies.
This fixes the situation that an event can be
a reply and in a thread. Before we have seen reply
as an relationshipType but this does conflict with
the concept of threads, where an event can be
of relationship type "thread" but also be a reply
with being a fallback or not.